refactor: 폴더 구조 관리
