+++ src/components/dialog/index.ts
@@ -1,0 +1,1 @@
+export { OverlapDialog } from './OverlapDialog';
