+++ src/components/event/EventList.tsx
@@ -9,7 +9,7 @@ import {
   Typography,
 } from '@mui/material';
 
-import { Event } from '../types';
+import { Event } from '../../types';
 
 interface EventListProps {
   searchTerm: string;
