+++ src/components/event/index.ts
@@ -1,0 +1,2 @@
+export { EventFormComponent } from './EventForm';
+export { EventList } from './EventList';
