+++ src/components/index.ts
@@ -1,0 +1,11 @@
+// Event related components
+export * from './event';
+
+// Calendar related components
+export * from './calendar';
+
+// Notification related components
+export * from './notification';
+
+// Dialog related components
+export * from './dialog';
